Initial video float
on the add video section made it so the fourm will be on the left side of the page and the video will appear in the middle right side. the problem right now is that if you make the page smaller the video will eventually overlap with the form. needs some fine tuning